+++ R/nestedMethods.R
@@ -54,7 +54,7 @@
 #' @keywords regression
 #' @examples
 #' # define continuation dichotomies for level of education
-#' cont.dichots <- continuationLogits(c("l.t.highschool",
+#' cont.dichots <- continuationLogits(c("<highschool",
 #'                                      "highschool",
 #'                                      "college",
 #'                                      "graduate"))