+++ R/predict.nestedLogit.R
@@ -84,7 +84,7 @@
 #' @keywords regression
 #' @examples
 #' # define continuation dichotomies for level of education
-#' cont.dichots <- continuationLogits(c("l.t.highschool",
+#' cont.dichots <- continuationLogits(c("<highschool",
 #'                                      "highschool",
 #'                                      "college",
 #'                                      "graduate"))
@@ -98,7 +98,7 @@
 #' predict(m)
 #'
 #' # predicted probabilities at specific values of predictors
-#' new <- expand.grid(parentdeg=c("l.t.highschool",  "highschool",
+#' new <- expand.grid(parentdeg=c("<highschool",  "highschool",
 #'                                "college", "graduate"),
 #'                    year=c(1972, 2016))
 #'