+++ NEWS.md
@@ -14,7 +14,7 @@ This is a major release of the package adding considerable functionality for plo
 * Added vignette: "Other Examples of Nested Logit Models"
 * In the `GSS` data, "l.t.highschool" changed to "<highschool"
 * Fixed problem with `lty` not passed to `matplot()` in `plot.nestedLogit()`
-* In `plot.nestedLogit()`, added `label.col` arg (defaulting to the value of `col` in the call), to control the color of the curve label
+* In `plot.nestedLogit()`, added `label.col` argument (defaulting to the value of `col` in the call), to control the color of the curve label
 
 ## nestedLogit 0.3.4